+++ packages/@expo/cli/src/start/platforms/android/__tests__/gradle-test.ts
@@ -1,10 +1,8 @@
 import spawnAsync from '@expo/spawn-async';
 
+import { asMock } from '../../../../__tests__/asMock';
 import { AbortCommandError } from '../../../../utils/errors';
 import { assembleAsync, formatGradleArguments, installAsync, spawnGradleAsync } from '../gradle';
-
-const asMock = <T extends (...args: any[]) => any>(fn: T): jest.MockedFunction<T> =>
-  fn as jest.MockedFunction<T>;
 
 jest.mock('../../../../utils/env', () => ({
   env: {