+++ packages/@expo/cli/CHANGELOG.md
@@ -13,6 +13,8 @@
 
 ### 💡 Others
 
+- deduplicate asMock helper function. ([#17294](https://github.com/expo/expo/pull/17294) by [@wschurman](https://github.com/wschurman))
+
 ## 0.1.3 — 2022-04-28
 
 ### 🐛 Bug fixes